refactor: moved Adapters to individual files
